move loadPrism
